replaced models.py with models folder and subfiles for each model
