+++ catalogue/models/__init__.py
@@ -1,0 +1,9 @@
+from .video import Video
+from .channel import Channel
+from .series import Series
+from .ministry import Ministry
+from .speaker import Speaker
+from .topic import Topic
+
+# NB will need to import models from the other tables
+# when the classes for these have been created
