+++ catalogue/models/label.py
@@ -1,0 +1,2 @@
+from django.db import models
+from django.urls import reverse # generate urls by reversing url pattern
